requirements.txt
# System dependency required for OCR (image indexing):
#   macOS:  brew install tesseract
#   Linux:  sudo apt install tesseract-ocr
chromadb
pdfplumber==0.11.9
pytesseract==0.3.13
Pillow==12.2.0
